Rename "Hardware" category to "Official Hardware"
This category serves as a container for dedicated subcategories for the individual products sold by the Arduino company.

The forum contains several other categories used for discussion of subjects related to hardware not specific to an
official product.

The previous ambiguous "Hardware" category name did not clearly communicate the scope of this category and this
ambiguity might have made it difficult for users to determine where to create or search for topics related to other
hardware. The new name clearly communicates the scope of the category.